requirements.txt
praw==7.8.1
requests==2.32.5
python-dotenv==1.1.1
jinja2==3.1.6
openai==1.100.2